+++ src/components/StatementBrowser/ValueForm/ValueForm.js
@@ -58,8 +58,7 @@ const ValueForm = props => {
     const [suggestionType, setSuggestionType] = useState(null);
 
     /* Select component reference can be used to check if menu is opened */
-    const isMenuOpen = () => autocompleteInputRef.current.state.menuIsOpen && autocompleteInputRef.current.props.options.length > 0;
-
+    const isMenuOpen = () => autocompleteInputRef.current.props.menuIsOpen && autocompleteInputRef.current.props.options.length > 0;
     const onSubmit = () => {
         let error;
         if (schema) {
@@ -139,7 +138,7 @@ const ValueForm = props => {
                         entityType={entityType}
                         excludeClasses={
                             entityType === ENTITIES.RESOURCE && !valueClass
-                                ? `${CLASSES.CONTRIBUTION},${CLASSES.PROBLEM},${CLASSES.TEMPLATE},${CLASSES.TEMPLATE_COMPONENT},${CLASSES.PAPER_DELETED},${CLASSES.CONTRIBUTION_DELETED}`
+                                ? `${CLASSES.CONTRIBUTION},${CLASSES.PROBLEM},${CLASSES.TEMPLATE},${CLASSES.TEMPLATE_COMPONENT},${CLASSES.PAPER_DELETED},${CLASSES.CONTRIBUTION_DELETED},${CLASSES.EXTERNAL}`
                                 : null
                         }
                         optionsClass={entityType === ENTITIES.RESOURCE && valueClass ? valueClass.id : undefined}
@@ -148,7 +147,7 @@ const ValueForm = props => {
                             handleAddValue(entityType, { ...i, label: i.value, selected: true });
                             props.setShowAddValue?.(false);
                         }}
-                        ols={entityType === ENTITIES.CLASS}
+                        ols={!valueClass}
                         onInput={(e, value) => setInputValue(e ? e.target.value : value)}
                         value={inputValue}
                         additionalData={newResources}
@@ -162,7 +161,7 @@ const ValueForm = props => {
                                 // escape
                                 props.setShowAddValue?.(false);
                             } else if (e.keyCode === 13 && !isMenuOpen()) {
-                                handleAddValue(entityType, { label: inputValue, selected: true });
+                                inputValue && handleAddValue(entityType, { label: inputValue, selected: false });
                                 props.setShowAddValue?.(false);
                             }
                         }}